chore: finalizado processo sde autenticação do usuário no app
